+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## UNRELEASED
 
+### ENHANCEMENTS
+
+* Bootstrap Yorc with a Vault instance ([GH-282](https://github.com/ystia/yorc/issues/282))
+
 ### BUG FIXES
 
 * Deployment fails on error "socket: too many open files" ([GH-334](https://github.com/ystia/yorc/issues/334))